feat: 新增接口

+++ src/main/java/com/antiy/asset/controller/AssetGroupController.java
@@ -8,17 +8,13 @@ import io.swagger.annotations.*;
 import com.antiy.common.base.ActionResponse;
 
 import javax.annotation.Resource;
-import com.antiy.common.base.QueryCondition;
-import com.antiy.common.utils.ParamterExceptionUtils;
 
 import com.antiy.asset.service.IAssetGroupService;
 import com.antiy.asset.entity.vo.request.AssetGroupRequest;
 import com.antiy.asset.entity.vo.query.AssetGroupQuery;
 
 
-
 /**
- *
  * @author zhangyajun
  * @since 2018-12-29
  */
@@ -42,9 +38,13 @@ public class AssetGroupController {
             @ApiResponse(code = 200, message = "OK", response = ActionResponse.class, responseContainer = "actionResponse"),
     })
     @RequestMapping(value = "/save/single", method = RequestMethod.POST)
-    public ActionResponse saveSingle(@RequestBody @ApiParam(value = "assetGroup")AssetGroupRequest assetGroup)throws Exception{
-        iAssetGroupService.saveAssetGroup(assetGroup);
-        return ActionResponse.success();
+    public ActionResponse saveSingle(@RequestBody @ApiParam(value = "assetGroup") AssetGroupRequest assetGroup) throws Exception {
+        System.out.println(assetGroup);
+        Boolean success = iAssetGroupService.saveAssetGroup(assetGroup) > 0;
+        if (success) {
+            return ActionResponse.success();
+        }
+        return ActionResponse.fail(RespBasicCode.ERROR, "保存数据失败");
     }
 
     /**
@@ -59,12 +59,19 @@ public class AssetGroupController {
     })
     @RequestMapping(value = "/update/single", method = RequestMethod.PUT)
     public ActionResponse updateSingle(@RequestBody @ApiParam(value = "assetGroup") AssetGroupRequest assetGroup) throws Exception {
-        iAssetGroupService.updateAssetGroup(assetGroup);
-        return ActionResponse.success();
+        Integer integer = iAssetGroupService.updateAssetGroup(assetGroup);
+        Boolean success = integer > 0;
+        if (success) {
+            ActionResponse.success();
+        } else {
+            return ActionResponse.fail(RespBasicCode.ERROR, "更新数据失败");
+        }
+        return null;
     }
 
     /**
      * 批量查询
+     *
      * @param assetGroup
      * @return actionResponse
      */
@@ -83,13 +90,13 @@ public class AssetGroupController {
      * @param id
      * @return actionResponse
      */
-    @ApiOperation(value = "通过ID查询", notes = "主键封装对象")
+    @ApiOperation(value = "通过id查询接口", notes = "主键封装对象")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "OK", response = ActionResponse.class, responseContainer = "actionResponse"),
     })
     @RequestMapping(value = "/query/{id}", method = RequestMethod.GET)
     public ActionResponse queryById(@PathVariable("id") @RequestBody @ApiParam(value = "id") Integer id) throws Exception {
-         return ActionResponse.success(iAssetGroupService.getById(id));
+        return ActionResponse.success(iAssetGroupService.getById(id));
     }
 
     /**
@@ -104,7 +111,11 @@ public class AssetGroupController {
     })
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
     public ActionResponse deleteById(@PathVariable @RequestBody @ApiParam(value = "query") Integer id) throws Exception {
-        return ActionResponse.success(iAssetGroupService.deleteById(id));
+        Boolean success = iAssetGroupService.deleteById(id) > 0;
+        if (success) {
+            return ActionResponse.success(iAssetGroupService.deleteById(id));
+        }
+        return ActionResponse.fail(RespBasicCode.ERROR, "删除数据失败");
     }
 }
